fate: add skip_clean option
This is useful if one wants to inspect build artifacts after running
fate.sh script.

+++ tests/fate.sh
@@ -95,7 +95,7 @@ fate()(
 )
 
 clean(){
-    rm -rf ${build} ${inst}
+    test "$skip_clean" = "yes" || rm -rf ${build} ${inst}
 }
 
 report(){
